# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FASE 2 API",
        "description": "FASE 2 Swagger OpenApi",
        "contact": {
            "email": "eaguilar.arrezola@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/cat_type_sales/paginate": {
            "get": {
                "tags": [
                    "cat_type_sales"
                ],
                "summary": "Get cat type sales per paginate",
                "operationId": "e96f568a0326401cd65b02198d309197",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/cat_type_sales": {
            "get": {
                "tags": [
                    "cat_type_sales"
                ],
                "summary": "Get all cat type sales",
                "operationId": "48d6b0470afaaee92c668778d71c1f18",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "cat_type_sales"
                ],
                "summary": "Get cat type sales",
                "operationId": "b130b562ad8f83a6ad176c53c154eba9",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/cat_type_sales/{id}": {
            "get": {
                "tags": [
                    "cat_type_sales"
                ],
                "summary": "Get cat type sales",
                "operationId": "ec8d056139cbfc9f699c7d1332bc8a9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "cat_type_sales"
                ],
                "summary": "Update cat type sales",
                "operationId": "ffae3fffcb2c8d8fcdf3ce64e33915d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "cat_type_sales"
                ],
                "summary": "Delete cat type sales",
                "operationId": "69b89d214c4e8c9bd6edd3f460a979ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients": {
            "get": {
                "tags": [
                    "clients"
                ],
                "summary": "Get all clients",
                "operationId": "4be9bbde54b325142519d9b8ae98b3b4",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "clients"
                ],
                "summary": "Add client",
                "operationId": "044543d7eec2742cf73c6dd72fb5f4af",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "clients"
                ],
                "summary": "Get client",
                "operationId": "b8445c760c565bf3eb59e44306e694a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "clients"
                ],
                "summary": "Update client",
                "operationId": "7d0a5ae9e306858d30e04fc50265d80c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "clients"
                ],
                "summary": "Delete client",
                "operationId": "8720346de08f1e46fe28a2a7179d23c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/paginate": {
            "post": {
                "tags": [
                    "clients"
                ],
                "summary": "Get clients per paginate",
                "operationId": "aac872b53af34ec165dd32739b18dda5",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}